added new package "application" for application.fxml, Main, and controller. Renamed TestClient class to Client.
